backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose
PyJWT
email-validator
google-genai
